+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `MockFn::debug_inputs` now returns `Vec<Option<String>>`.
 - `macro_api::Evalution` is now `#[non_exhaustive]`.
 - `macro_api::Evalution::Skipped` is not called `Unmocked`.
+- `MockFn::NAME` replaced by `fn info() -> MockFnInfo`.
 
 ## [0.4.12] - 2023-03-29
 ### Fixed